requirements.txt
streamlit==1.35.0
tensorflow==2.13.0
pillow==9.5.0
numpy==1.23.5